requirements.txt
#faker==14.2.0
parsedatetime==2.6
plotly
dataframe_image==0.2.3
hopsworks==3.7.0
hsml
joblib
matplotlib==3.6.3
pandas==1.5.1
scikit-learn==1.4.0
seaborn==0.13.2
python-dotenv
requests
alpha_vantage
textblob
great_expectations==0.18.12
#prophet
tensorflow
pandas_market_calendars
streamlit
streamlit_folium
asyncio
nest_asyncio
opencv-python-headless
easyocr
Pillow
python-dotenv
